Finally! I've managed to figure out what was wrong with the code for laying out the recipes for viewing!  From time to time, recipe elements would overflow into others, or the background area wouldn't match the content area.
The problem was in calculating the height of an element according to its width, and now the method for determining this has been fixed to calculate the precise height no matter what the contents or width.  Now the recipe view should be as expected and flawless ;)

Consequently, this also fixes the same issue in the HTML exporter.

svn path=/trunk/kdeextragear-3/krecipes/; revision=296445