Big feature here: PostgreSQL backend!
It appears to be fully functional and complete.  Simply supply the server parameters and set DBType="PostgreSQL" in the krecipesrc config file and if the server is running and permissions are set, you are good to go.

What is left is to implement support in the wizard.

Note that in the process I've re-organized the database backend code to make supporting more backends very easy.

-And on a minor note, when storing recipe photos, write it directly into a QByteArray instead of writing a JPG to the file system and reading that into a QByteArray

svn path=/trunk/kdeextragear-3/krecipes/; revision=351281