-Import recipes from any Krecipes database backend.  The behind-the-scenes stuff is done, but a GUI for choosing database parameters is still needed.  Also I still need to fix a crash when the database chosen is the same as the current database (or prevent that from happening). -Update configure scripts to recognize PostgreSQL and to allow disabling particular db backends
internal - new function: RecipeDB::createDatabase() which is used to get a database connection.  It really simplifies things, since now we don't need the series of if-else/#ifdef statements every time we need a new database.  We also don't have to deal with individual backends, such as when new ones are added.

svn path=/trunk/kdeextragear-3/krecipes/; revision=362180