List views of database elements (ingredients, units, etc.) are now loaded in set increments.  Since large databases can have tens of thousands of ingredients/prep methods/etc., we can now only load subsections of the total elements to save memory and thus make larger databases possible.  For example, only 5,000 ingredients may be loaded at once when there are 50,000 total.  The user can browse ingredients 1-5000, then 5001-10000, etc.
Currently, the default is still to load them all (although it can be set manually in the config file).  Also, the only way to load the next increment is via keyboard shortcuts.

TODO: Check the upper limit of a range that is being loaded.  Right now, it will query the database for elements out of range of what is in the database (It may ask for the ingredients 500-1000 when there are only 200 in the database).

svn path=/trunk/kdeextragear-3/krecipes/; revision=390822