GUI: In the recipe list, categories now have an icon associated with it (it is a folder icon now, but may change... I'm open to any thoughts on this).  Also, the "+" to expand each category is not visible until the item is clicked (or a search if performed), reason being that the recipes within the category don't exist in the list until they are requested.  However, for the user, this shouldn't need to even be mentioned since they'll see the folder icon and think to click it to see what items are in it.  For the handbook, I'd think that only screenshots would need to be updated (Find/Edit and Shopping List pages).
What is happening on the inside is that only the top-level categories are loaded initially.  When "space" or "return" is pressed on the keyboard or when the category item is executed (meaning single- or double-clicked depending on global KDE settings), Krecipes queries the database and loads the list with the recipes in the selected category.

This means a lot less to load at startup and at various times througout the use of Krecipes.  I've been working with a large database and there were several bottlenecks dealing with this.  In particular, the dialog to select a category for a recipe took about 20 seconds to load, and now appears instantaneously.  Also, some other slow areas were found and fixed, such as when inputting a recipe.  I've used the main features of Krecipes on this huge database and there is no longer a sluggishness like there has been.

And finally, the current startup time: 3 seconds.  On my system, that's practically the same time as an empty database :-)

Important note: Startup speed is really slowed down when recipes are not categorized (With 600 uncategorized recipes, it takes 15 seconds to load the uncategorized recipes and less than one to load the rest).  Somewhere, we should let the users know that and encourage them to categorize all their recipes.  Maybe we need a "Performance Tips" section in the Handbook... just a thought.

Anyway, this should be last big commit until 0.8.  It very likely introduced some bugs, but after this, no big changes will be made and things will only get more stable.

I hope I got everything ;-)

svn path=/trunk/extragear/utils/krecipes/; revision=417222