This is the beginning of a new system for allowing improved and more flexible printing support.  The idea is to allow printing to any medium, such as index cards or in a precise way as to fit into a cookbook page formatted in a particular manner.
What needs to change is that there must be a layout for printing and a 
layout for viewing.  This way, a layout can be chosen for regular 
viewing, and then different layouts can be created for printing to 
different mediums.  Printing to a different medium will be as simple as 
selecting a different layout.  When File->Print is selected, the print 
layout will be used.

The layout for printing will be based on physical dimensions.  One 
obstacle that must be overcome is that when limiting the view to a 
particular set of dimensions, the view cannot expand to accomodate it.  
For example, if the instructions of a recipe do not fit within the space 
given, the print view should not expand as it may go beyond the physical
dimensions.  To handle this, a "Overflow" option has been added to the 
views.  The text overflow can either expand vertically or it will 
"Shrink to Fit" by decreasing the font size.  The default print layout 
will use the "Shrink to Fit" option.

Keep in mind, many of these changes have yet to be completed and this is 
preliminary work to make this possible.

And then, one other tiny change:
 -When right-clicking an item in the Page Setup and opening the 
 Alignment menu, the current setting for the alignment will be turned on
 

svn path=/trunk/extragear/utils/krecipes/; revision=431370