More fixes from the 0.8 branch.  This includes another search performance increase, this time in the advanced search.
Also, there's some work on increasing performance when loading the recipe edit page.  This can be done by using a progressive loading of 
ingredients for the auto-completion based on a QTimer.start(0,false) to load a portion of ingredients at a time without holding up the GUI 
thread.  The code is in place, but not put to use.

svn path=/trunk/extragear/utils/krecipes/; revision=435766