Support for database backup/restore is coming.  While you can currently export all the recipes in your database, the backup feature will make an exact copy of the database as-is, including nutrient information and unit conversion.
The created backup file can only be restored into the same database backend from 
which it came (A backup of SQLite cannot be restored into a MySQL database).  However, 
a backup from one version of a database is typically compatible with another 
version of the same database (SQLite 2 and 3 *are* interchangable, which is 
useful in upgrading from SQLite 2 to 3).

Backups of each database rely on the following binaries being available:
	SQLite: sqlite or sqlite3 (depending on which version Krecipes was 
					compiled with)
	MySQL: mysqldump
	PostgreSQL: pg_dump

As some might have guess from this information, the backup file is actually a 
database dump.

This new code allows Krecipes to create the backup, and code for restoring 
it will follow.  Access to creating the backup is only available via DCOP 
functions on the command-line.


svn path=/trunk/extragear/utils/krecipes/; revision=450794