Add an infrastructure for creating tests for the importers.  The idea is to be able to run "make check" and test programs will be created for each importer.  It will test that a given built-in recipe matches up with an identical recipe, formatted in the given recipe format, after it is imported.
This commit includes a test for the MealMaster format, and the tests 
for the other formats will follow.

To allow creating a program that uses only the Krecipes datablocks and 
importer data, I removed all instances of dialog code from the 
importers and moved it to a more appropriate place in the source tree.  
This also allows for a cleaner source tree and structure, and clearer 
separation of the different parts of Krecipes.

svn path=/trunk/extragear/utils/krecipes/; revision=457199