New feature: There is a dialog to paste your ingredients into.  You paste the ingredients into the text edit, Krecipes will parse the ingredients, and then you may edit the ingredients before adding them to the recipe (in case any were incorrectly parsed).
This is accessed through the clipboard icon below the "Remove 
Ingredient" button in the recipe edit dialog.

The parsing algorithm is pretty simple.  It looks for ingredients in the 
format of: [amount] [unit] [ingredient](; or ,)[prep methods]

This design works for the popular recipe sites I've checked, such as 
Food Network, AllRecipes.com, and Epicurious.  I'd hope that this covers 
mostly all cases when copying ingredients from an external source (but 
please prove me wrong if I am ;-).  I might give an option to specify 
what separates ingredients from prep methods, but from all the sites I've seen, the ';' and ',' are by far 
the most common.

Each line entered is considered an ingredients.  For each line, the 
parsing goes from left to right.  It tests the beginning for an 
amount, which can be a decimal, mixed number, or range.  If no amount is 
found, it tests that it is a unit.  Whether or not this text is 
considered a unit is determined by loading the units already in the 
database and comparing it to that.  If it doesn't match an existing 
unit, it will then be considered an ingredient.  Then, of what 
is left to parse, anything after a "," or ";" is considered the prep 
method(s).

Feel free to comment (or complain) about the design ;-)

Also:
 -New half-complete NYC importer test, and to go with it are some fixes 
  to the importer dealing with yields
 -Fix the photo from the last editted recipe getting picked up by the next

svn path=/trunk/extragear/utils/krecipes/; revision=459483