GUI: Add a new action, "Copy to Clipboard".  This is accessible via the Edit menu or by right-clicking a recipe.
Also, make the action in the menubar match up better with what the user would expect.  
The menubar actions, like "Export" and "Edit Recipe", were linked only with the basic
search tab.  That means, if you selected from the basic search tab, switched to the advanced search
and selected a recipe there, the menubar action would be for the recipe selected back at the basic
search tab.

svn path=/trunk/extragear/utils/krecipes/; revision=462022