Why, oh why, wasn't this done sooner... A separate codebase for the SQLite backend is no more.  Instead, it now works like the MySQL and PostgreSQL backends, which use the QSql interface.  Since the Qt plugin only supports SQLite 2.x, I've written my own that supports both (It lacks some features of Qt's plugin, but supports all that Krecipes uses).  Finally there is one class which handles all database communication, rather than duplicating the effort in two places.
Also, the photo data was stored differently for each backend.  Now, they all store the photo 
in the base64 encoding (which PostgreSQL already used).  The database will be 
automatically ported to convert all existing photos to this format.

This changes quite a bit of things, and needs lots of testing.  Be on the lookout for weird 
behavior.

svn path=/trunk/extragear/utils/krecipes/; revision=463765