It's about time I implemented this: allow selecting which ingredient properties to display.
This can be done via Data->Properties and checking/unchecking the tickbox next to a 
property.

It is important to note that properties to show are stored in the user configuration 
file, and NOT the database.  This means users can select which properties without 
affecting other users who are potentially sharing the same database (i.e. a remote 
database).

svn path=/trunk/extragear/utils/krecipes/; revision=547455