Implement ingredient substitutes.  Any ingredient can now have multiple substitutes, (i.e. butter or margarine).
This affects a lot of the code, so there's bound to be some issues that I need to take care 
of.  Other features still need to be updated to work with ingredient substitutes such as 
import/export and the advanced search (substitutes are currently ignored).

Also, update to the USDA SR18.

svn path=/trunk/extragear/utils/krecipes/; revision=553783