Users can specify how much of an ingredient is available in the ingredient matcher.  If no amount is specified, it is assumed that enough is available.
This feature is made possible by the new conversion routines ;-), which are used to 
convert how much of an ingredient is in a recipe to the units in the amount that is 
available.

For this feature, there was a change to how recipes are loaded... if anything seems funny 
relating to that, this is the culprit.

svn path=/trunk/extragear/utils/krecipes/; revision=563313