Add a property status indicator in the ingredient tab of the recipe input.  This currently tells whether or not all nutrient information is known for each ingredient in the recipe.
The GUI isn't done yet, so no, that giant LED isn't going to stay that big.  Also, the "Details..." 
button doesn't do anything yet.  At the moment, the information that this button will give is just 
spit out onto stdout (the console).  You'll notice this information is prefixed with "NUTRIENT STATUS:".

So this should be fully-functional, but the GUI will be finished later.

svn path=/trunk/extragear/utils/krecipes/; revision=564767