The "Details..." button works now ;-)
In the Ingredient tab of recipe input, the "Details..." button now pops up a non-modal dialog that 
displays what information is missing for complete nutrient analysis.

The dialog is updated as ingredients are added and removed and is automatically closed if nutrient 
information is complete.

The status messages should be complete information-wise, but might be tweaked or re-worded for 
clarity and aesthetics (I do need to make the 'can't convert from any of..." message clearer).

svn path=/trunk/extragear/utils/krecipes/; revision=565710