In the recipe import dialog, keep identical recipe items in sync.
For example, the recipe "Candy Cane Cookies" may be under both the "Christmas" and 
"Cookie" categories.  The recipe will therefore appear twice in the list of recipes 
to import.  Now unselecting the entry under one category also unselects every other 
entry as well.

It was possible before to unselect a recipe and it would still be imported because 
it existed selected somewhere else in the list... that was rather annoying, though 
subtle, and I'm glad it's fixed ;-) (that's another annoyance that couldn't be in 1.0).

svn path=/trunk/extragear/utils/krecipes/; revision=566075