Update ingredient weights to also include how the ingredient is prepared.  CAUTION: For updating the database structure, it just wipes out any existing weight info!
There is also a yellow status for approximations, and if a prep method is used, but there isn't a weight for it, the status will be 
yellow and it will use an entry without a prep method.  Note that the status will be red if there is missing info and approximations, 
and will be yellow if there is no missing information (red status) but just approximations.

Most of the work is done, but it needs some tweaking (when editting, the prep method combo box is hidden unless you resize the Per 
Amount column)

No further work has been done on loading USDA data, so imported USDA data still won't automatically use prep methods.

svn path=/trunk/extragear/utils/krecipes/; revision=566373