-Clean up the USDA weight portion of the import.
What it does now is looks at a list of known units and prep methods.  
If the unit of an entry is in the list, it'll create a weight entry 
using it.  This way, an arbitrary number of units won't end up being 
created, only those in the list can be created from the import.

The unit is considered the part before a comma, the rest is the prep method.  
Once the unit is accepted, it checks that the prep method matches one of 
the acceptable entries.  If it does, it adds the weight using that prep 
method.  If not, it stills adds the entry, but without whatever extra 
info the USDA entry held (which most often isn't of importance).

-Fix some indexes in the SQLite database.  Some weren't being created 
which would slow things down a bit.

svn path=/trunk/extragear/utils/krecipes/; revision=570167