GUI: The dialog that warns that some items are going to be deleted if another item is deleted (i.e. a recipe will be deleted because it uses a unit being deleted) has been cleaned up both visibly and internally.
Visibly, the items being deleted are displayed in a listbox rather than 
a listview (the visible difference is the listbox doesn't have a column 
header)

Internally, the DependanciesDialog now supports arbitrary lists of 
dependancies, not just for recipes and ingredients.

Also, I've made use of this change to allow displaying ingredient 
weights that have to be deleted when a unit they depend on is deleted.  
This is much better than silently deleting the data ;-)

svn path=/trunk/extragear/utils/krecipes/; revision=572323