Localization update to allow *all* base data to be localized.
Instead of putting all this base data inside of a header file in the 
form of an array, they are loaded from a data file at runtime based on 
locale.  There are four files to translate: property-data-LANG.txt, 
ingredient-data-LANG.txt, unit-data-LANG.txt, and prep-data-LANG.txt.

Also, the user gets a message if they ask to import nutrient data and 
there is no localized data for their language.  They can continue, using 
English data, or decide not to do the import.

svn path=/trunk/extragear/utils/krecipes/; revision=582244