This commit fixes a wrong use of QSqlDatabase::removeDatabase(...) according to the Qt documentation, we need to delete the related QSqlQuery and QSqlDatabase objects BEFORE calling removeDatabase, otherwise a resource leak will occur.
Before this revision I was experiencing sometimes crashes with the mysql
backend on exiting the program. On each exeution I was getting very different
backtraces, i.e. a freaking unreproducible bug. Hopefully this commit fixes the
issue, or, at least, works for me.

Finally, before this revision I was getting this warning in the program stdout:
connection 'connection1' is still in use, all queries will cease to work
which puts me on the road to fix this bug.


svn path=/trunk/extragear/utils/krecipes/; revision=982119