Create a QValidator for MixedNumber.
Also I have added an incomplete mixednumber test. I needed it to solve a bug
which I introduced when I was working on this commit.

Furthermore, I removed the default value for locale_aware, in my opinion it's
too prone to mistakes.

This new validator isn't used yet, but it will be useful in the future.