Comment out the encoding failing test.
If I'm not mistaken, the strings stored in database are being escaped in
QSqlRecipeDB::escapeAndEncode() and they are in UTF-8 encoded as latin1,
so some UTF-8 strings just can't be represented.

I guess the idea was being able to store any string wihtout caring about the
DBMS encoding configuration, however that's wrong.

I have been aware of this Krecipes enconding horror for a long time, but I
didn't have time yet to fix it, since fixing it would mean bumping the database
schema version number and re-encoding all the existing strings in the existing
databases. This is noted in the TODO file.

CCMAIL: groot@kde.org