Revert "Clean up the author tests case and make them work"
This reverts commit 60645572c72a94320be95f5ed2b9ce71ddff67fe.

The authors test was actually a mockup of future tests to check that Krecipes'
SQL queries can be executed in the three different backends, It's needed to
test the queries in SQLite, MySQL and PostgreSQL (and not just one of them)
because:

1. There are some small differences between SQLite, MySQL and PostgreSQL (a
   query can work, for instance, on MySQL but not on SQLite).

2. Due to the differences between them they are handled sometimes in a different
   way in Krecipes (and Qt SQL backends), so there are bugs affecting one
   backend but not the others, for instance some days ago this program was
   failing to store photos *only* in SQLite.

The problem now is how to make the tests pass in KDE's jenkins: the MySQL and
PostgreSQL tests can't be executed there because they work on the expectation
of having a configured server. Therefore, obviously, the right way to proceeed
is disabling the MySQL and PostgreSQL tests by default, not removing them.